requirements.txt
# requirements.txt
selenium
openai==0.28
pyyaml
webdriver-manager